Refactoring code into separate files.
