+++ Sources/LogBookView.swift
@@ -1,6 +1,5 @@
 import SwiftUI
 
-// MARK: - LogBookView
 public struct LogBookView: View {
 
     let logBook: LogBook = .sharedInstance
@@ -56,123 +55,6 @@ public struct LogBookView: View {
                 subsystems = await logBook.subsystems
             }
         }
-    }
-}
-
-// MARK: - LogTypeTag
-struct LogTypeTag: View {
-    let type: LogType
-
-    private var label: String {
-        switch type {
-        case .error: return "ERRR"
-        case .warning: return "WARN"
-        default: return "INFO"
-        }
-    }
-
-    private var color: Color {
-        switch type {
-        case .error: return .red
-        case .warning: return .yellow
-        default: return .green
-        }
-    }
-
-    var body: some View {
-        Text(label)
-            .font(.system(size: 8, weight: .bold, design: .monospaced))
-            .foregroundStyle(.black)
-            .padding(.horizontal, 4)
-            .padding(.vertical, 2)
-            .background(Capsule().fill(color).opacity(0.65))
-    }
-}
-
-// MARK: - LogRowView
-struct LogRowView: View {
-    let log: LogBookItem
-
-    var body: some View {
-        LazyVStack(alignment: .leading, spacing: 12) {
-            Text(log.message)
-                .font(.system(size: 12, weight: .regular, design: .monospaced))
-                .textSelection(.enabled)
-            HStack(spacing: 5) {
-                LogTypeTag(type: log.type)
-                Text(log.subsystem.uppercased())
-                    .foregroundStyle(.secondary)
-                    .font(.system(size: 10, weight: .bold, design: .monospaced))
-            }
-        }
-    }
-}
-
-// MARK: - LogListHeaderView
-struct LogListHeaderView: View {
-    let onClear: () -> Void
-
-    var body: some View {
-        HStack {
-            Text("Latest logs appear at the top.")
-            Spacer()
-            Button(action: onClear) {
-                Text("Clear all").bold()
-            }
-        }
-        .font(.system(size: 12, weight: .regular, design: .monospaced))
-        .foregroundStyle(.secondary)
-    }
-}
-
-// MARK: - LogListEmptyView
-struct LogListEmptyView: View {
-    var body: some View {
-        ContentUnavailableView(
-            "Can't find your logs?",
-            systemImage: "questionmark.message.fill",
-            description: Text("Please ensure Log Book is enabled using the toggle above.")
-                .font(.caption)
-        )
-        .listRowBackground(EmptyView())
-    }
-}
-
-// MARK: - LogBookHeaderView
-struct LogBookHeaderView: View {
-    @Binding var isEnabled: Bool
-    @Binding var selectedSubsystem: String
-    let subsystems: Set<String>
-
-    var body: some View {
-        HStack {
-            Image(systemName: "book.pages.fill")
-                .font(.system(size: 40, weight: .regular))
-            VStack(alignment: .leading, spacing: 5) {
-                Text("Log Book")
-                    .font(.system(size: 25, weight: .bold, design: .monospaced))
-                Text("View the latest Qalam logs.")
-                    .font(.system(size: 11, design: .monospaced))
-            }
-            Spacer()
-            VStack(spacing: 0) {
-                Toggle(isOn: $isEnabled) { }
-                    .frame(width: 10)
-                    .scaleEffect(0.9)
-                    .padding(.trailing)
-                    .tint(.indigo)
-                Picker("Subsystem", selection: $selectedSubsystem) {
-                    Text("All").tag("All")
-                    ForEach(Array(subsystems).sorted(), id: \.self) { subsystem in
-                        Text(subsystem).tag(subsystem)
-                    }
-                }
-                .pickerStyle(.automatic)
-                .tint(.primary)
-            }
-        }
-        .padding()
-        .dynamicTypeSize(.large)
     }
 }
